refactor: Merge NextPuyo and FieldPuyo
