+++ src/components/PuyoBoard.module.css
@@ -4,6 +4,18 @@
   border-radius: 2px;
 }
 
+.pointer {
+  cursor: pointer;
+}
+
+.crosshair {
+  cursor: crosshair;
+}
+
+.notAllowed {
+  cursor: not-allowed;
+}
+
 .boardBackground {
   stroke: none;
   fill: #000;
@@ -29,14 +41,6 @@
   fill: rgba(255, 255, 255, 0.5);
 }
 
-.pointer {
-  cursor: pointer;
-}
-
-.crosshair {
-  cursor: crosshair;
-}
-
-.notAllowed {
-  cursor: not-allowed;
+.puyo {
+  transition: transform 200ms ease;
 }
