+++ src/components/PuyoBoard.tsx
@@ -149,9 +149,9 @@ const PuyoBoard: React.FC = React.memo(() => {
       <BoardFrame />
       <GridLines />
       <BoostAreaView coordSetList={boostAreaCoordSetList} />
-      <g transform={`translate(${fw} ${fw})`}>
+      <g key="innerFrame" transform={`translate(${fw} ${fw})`}>
         <PuyoMatrix nextPuyos={nextPuyos} field={field} />
-        <g>
+        <g key="coords">
           {optimalTraceCoords?.map((coord, i) => (
             <OptimalTrace key={String(i)} x={coord.x} y={coord.y} />
           ))}
